ALL OF IT
NOT ONCE did I think to use a version control software, largely because this was a personal project of mine, but also because I forgot to. However, I now see how this program may be useful to others, and so now I shall let them have it. Enjoy! Made in 10 days by </>